requirements.txt
aniso8601==8.0.0
attrs==19.3.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.0
Flask-JWT==0.3.2
Flask-Mail==0.9.0
flask-restx==0.2.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
ldap3==2.9.1
MarkupSafe==1.1.1
numpy==1.22.3
pandas==1.4.2
PyJWT==1.4.2
pyrsistent==0.16.0
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zipp==3.1.0
psycopg2==2.9.4